Update WerewolfGame.java
Greetings. I'm William, @True_Incarnatus, a developer from the Kothar Labs. Today that I've completed the first prototype, and it might be one of the last few iterations. The Labs' kept quite busy prototyping the automated submarine which belongs to another project. In the sight of recent tests, I have a few issues to pinpoint. 
First, the language generating mechanism needs great improvement: Sentences spoken by game-bots are not very logical and understandable. Gamebots are reported to sometimes elevate and debase the same player at the same time and repeat themselves numerous times. A logical analysis system is to be introduced in later iterations to avoid repetitive or conflictive statements.
Second, strategy of gamebots playing civilians in game should be enhanced. According to our data, in 1000 trials, werewolves won for 993 times, while civilians only 7. As intensive analysis reveals, behavior of gamebot players during the day is rather random. They almost gamble around when they vote for banishment, which results in huge informational asymmetry of the two groups. It is obvious that werewolves, with much informational advantage, are much more likely to win the game. In later iterations, mechanisms should be introduced to make werewolves more "distinguishable" during days, like they somehow behave differently from civilians. Algorithm should also be improved to enable distinguishment of werewolves and civilians from speech.
Last but not least, vital parts of the code lack comments, as a result of hasty completion. Developers are now working on comments and documentations. Through later iterations they'll be introduced and refined. 
The good thing is, the first build is now running without reported runtime errors or exceptions during tests. That's the first milestone of the project achieved. Kothar Labs will make use of any room of improvements.